# lc-ssi-wc/openapi/swift-data-service.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baseline SWIFT Data Service",
    "version": "1.1.0-prototype",
    "description": "SR2026 prototype contract. BIC syntax follows ISO 9362; currency codes are supplied by the ISO 4217 Currency service. Licensed SwiftRef/RMA/SSI data remains a production prerequisite."
  },
  "servers": [
    {
      "url": "http://localhost:3100/api"
    }
  ],
...
  "paths": {
...
    "/v1/resolution-page-definitions/index": {
      "get": {
        "summary": "List governed resolution page definitions",
        "parameters": [
          {
            "name": "standardsRelease",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "businessDomain",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageParameterBusinessDomain"
            }
          }
        ],
        "responses": {
          "200": { "description": "Governed definition index" }
        }
      }
    },
    "/v1/resolution-page-definitions/execute": {
      "post": {
        "summary": "Execute a governed SSI resolution page submission",
        "description": "Executes only scenarios owned by SSI_FIELD_RESOLUTION_API with action RESOLVE_SSI. Upstream FIN validation scenarios use their separately governed endpoint.",
        "operationId": "executeResolutionPage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolutionPageSubmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Governed resolution result; payloadGenerated is false when no SSI route is produced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolutionPageExecutionResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or non-executable governed submission"
          },
          "404": { "description": "Governed definition or scenario not found" },
          "422": {
            "description": "Resolver rejected the submission or no eligible SSI exists"
          },
          "502": { "description": "Resolver returned an invalid response" }
        }
      }
    },
    "/v1/resolution-page-definitions/lookups/own-account-receivers": {
      "get": {
        "summary": "List governed Receiver Bank Services for an own-account settlement scenario",
        "operationId": "listOwnAccountReceivers",
        "parameters": [
          {
            "name": "scenarioId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "messageType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["MT202", "MT202COV", "MT205", "MT205COV"]
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Currency" }
          },
          {
            "name": "bookingEntity",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "valueDate",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Governed Receiver Bank Service choices",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageParameterLookupEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or incomplete governed lookup context"
          }
        }
      }
    },
    "/v1/resolution-page-definitions/lookups/nostro-accounts": {
      "get": {
        "summary": "List governed Nostro accounts for an own-account settlement role",
        "description": "Returns stable Nostro IDs with authoritative versions. The selected version is transported as a hidden companion value and is re-derived by the server when the resolution is executed.",
        "operationId": "listOwnAccountNostros",
        "parameters": [
          {
            "name": "scenarioId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "messageType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["MT202", "MT202COV", "MT205", "MT205COV"]
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": true,
            "schema": { "$ref": "#/components/schemas/Currency" }
          },
          {
            "name": "bookingEntity",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "valueDate",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "receiverBankServiceId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "ownDebitAccountId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "targetRole",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["OWN_DEBIT_ACCOUNT", "OWN_CREDIT_ACCOUNT"]
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Governed Nostro choices and lookup-derived companion version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageParameterLookupEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or incomplete governed lookup context"
          }
        }
      }
    },
    "/v1/resolution-page-definitions": {
      "get": {
        "summary": "Get one governed resolution page definition",
        "parameters": [
          {
            "name": "standardsRelease",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "messageFamily",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "messageType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["MT202", "MT202COV", "MT205", "MT205COV"]
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "enum": ["INCOMING", "OUTGOING"] }
          },
          {
            "name": "businessDomain",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PageParameterBusinessDomain"
            }
          },
          {
            "name": "businessScenarioId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "businessService",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["swift.cbprplus.04", "swift.cbprplus.cov.04"]
            }
          }
        ],
        "responses": {
          "200": { "description": "Versioned OAS-driven page definition" },
          "404": { "description": "No governed executable definition" }
        }
      }
    }
  },
  "components": {
...
    "schemas": {
      "PageParameterBusinessDomain": {
        "type": "string",
        "enum": ["TREASURY", "TRADE_FINANCE", "PAYMENT"]
      },
      "PageParameterLookupProvider": {
        "type": "string",
        "enum": ["BANK_SERVICE", "SSI_COUNTERPARTY", "NOSTRO_ACCOUNT"]
      },
...
      "PageParameterLookupResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["provider", "action", "displayValue"],
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/PageParameterLookupProvider"
          },
          "action": {
            "$ref": "#/components/schemas/PageParameterLookupProvider"
          },
          "bankServiceId": { "type": "string", "minLength": 1 },
          "bic": { "$ref": "#/components/schemas/Bic" },
          "nostroId": { "type": "string", "minLength": 1 },
          "version": { "type": "integer", "minimum": 1 },
          "accountServicerBic": { "$ref": "#/components/schemas/Bic" },
          "maskedAccountRef": { "type": "string", "minLength": 1 },
          "displayValue": { "type": "string", "minLength": 1 },
          "selectedRouteIdentity": {
            "$ref": "#/components/schemas/PageParameterSelectedRouteIdentity"
          }
        }
      },
      "PageParameterVersionedRecordIdentity": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "version"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "version": { "type": "integer", "minimum": 1 }
        }
      },
      "PageParameterSelectedRouteIdentity": {
        "type": "object",
        "additionalProperties": false,
        "description": "Exact indivisible SSI route selected from discovery. Every versioned component is revalidated before execution.",
        "required": ["routeId", "definitionId", "definitionVersion", "fixtureBindingId", "contextSha256", "ssi", "applicability", "nostro", "rma"],
        "properties": {
          "routeId": { "type": "string", "minLength": 1 },
          "definitionId": { "type": "string", "minLength": 1 },
          "definitionVersion": { "type": "string", "minLength": 1 },
          "fixtureBindingId": { "type": "string", "minLength": 1 },
          "contextSha256": { "type": "string", "pattern": "^[a-fA-F0-9]{64}$" },
          "ssi": { "$ref": "#/components/schemas/PageParameterVersionedRecordIdentity" },
          "applicability": { "$ref": "#/components/schemas/PageParameterVersionedRecordIdentity" },
          "nostro": { "$ref": "#/components/schemas/PageParameterVersionedRecordIdentity" },
          "rma": { "$ref": "#/components/schemas/PageParameterVersionedRecordIdentity" }
        }
      },
      "PageParameterLookupDefaultSelection": {
        "type": "object",
        "additionalProperties": false,
        "required": ["valueField", "value", "reasonCode"],
        "properties": {
          "valueField": {
            "type": "string",
            "enum": ["bankServiceId", "nostroId"]
          },
          "value": { "type": "string", "minLength": 1 },
          "reasonCode": {
            "type": "string",
            "enum": ["SINGLE_ELIGIBLE_DEFAULT", "GOVERNED_PRIORITY_DEFAULT"]
          },
          "companionValues": {
            "type": "object",
            "additionalProperties": false,
            "description": "Atomic hidden values belonging to the selected lookup record.",
            "properties": { "version": { "type": "integer", "minimum": 1 } }
          }
        }
      },
      "PageParameterLookupEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["provider", "action", "items", "eligibilitySnapshot"],
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/PageParameterLookupProvider"
          },
          "action": {
            "$ref": "#/components/schemas/PageParameterLookupProvider"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageParameterLookupResult"
            }
          },
          "eligibilitySnapshot": {
            "$ref": "#/components/schemas/PageParameterEligibilitySnapshot"
          },
          "defaultSelection": {
            "$ref": "#/components/schemas/PageParameterLookupDefaultSelection"
          }
        }
      },
      "PageParameterEligibilitySnapshot": {
        "type": "object",
        "additionalProperties": false,
        "description": "Opaque server-governed identity binding the displayed eligible choices to the subsequent resolve request.",
        "required": ["snapshotId", "contextSha256"],
        "properties": {
          "snapshotId": { "type": "string", "minLength": 1 },
          "contextSha256": { "type": "string", "pattern": "^[a-fA-F0-9]{64}$" },
          "snapshotIdentityMethod": { "type": "string", "minLength": 1 }
        }
      },
...
      "PageParameterCompatibleNvrOutcome": {
        "type": "string",
        "description": "Backward-compatible NVR outcome. N_A is accepted only for legacy page contracts.",
        "enum": ["PASS", "FAIL", "NOT_EVALUATED", "N_A"]
      },
...
      "ResolutionPageSubmission": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "definitionId",
          "definitionVersion",
          "scenarioId",
          "fixtureBindingId",
          "contractSha256",
          "eligibilitySnapshot",
          "selectedRouteIdentity",
          "values"
        ],
        "properties": {
          "definitionId": { "type": "string", "minLength": 1 },
          "definitionVersion": { "type": "string", "minLength": 1 },
          "scenarioId": { "type": "string", "minLength": 1 },
          "fixtureBindingId": { "type": "string", "minLength": 1 },
          "contractSha256": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{64}$"
          },
          "eligibilitySnapshot": {
            "$ref": "#/components/schemas/PageParameterEligibilitySnapshot"
          },
          "selectedRouteIdentity": {
            "$ref": "#/components/schemas/PageParameterSelectedRouteIdentity"
          },
          "values": {
            "type": "object",
            "description": "OAS-driven fieldId/value map. For Payment COV scenarios, the service derives 32A from governed typed date, currency and amount values, and merges server-governed 21, 119:COV, 121, Sequence B and controlled previous-message provenance from the selected scenario; clients must not invent or override them.",
            "properties": {
              "context.swift21NONE": {
                "type": "string",
                "minLength": 1,
                "readOnly": true
              },
              "context.swift119NONE": {
                "type": "string",
                "const": "COV",
                "readOnly": true
              },
              "context.swift121NONE": {
                "type": "string",
                "format": "uuid",
                "readOnly": true
              },
              "context.sequenceB50A": {
                "type": "string",
                "minLength": 1,
                "readOnly": true
              },
              "context.sequenceB59": {
                "type": "string",
                "minLength": 1,
                "readOnly": true
              },
              "context.previousMessageType": {
                "type": "string",
                "enum": ["MT202COV", "MT205COV", "EQUIVALENT_COVER"],
                "readOnly": true
              },
              "context.previousMessageArtifactSha256": {
                "type": "string",
                "pattern": "^[a-fA-F0-9]{64}$",
                "readOnly": true
              }
            },
            "additionalProperties": {
              "oneOf": [{ "type": "string" }, { "type": "boolean" }]
            }
          }
        }
      },
      "ResolutionPageGeneratedOutput": {
        "type": "object",
        "additionalProperties": false,
        "description": "A server-generated representation of one confirmed canonical SSI resolution. Clients render the document and must not translate between formats.",
        "required": [
          "outputId",
          "format",
          "label",
          "messageIdentity",
          "mediaType",
          "document"
        ],
        "properties": {
          "outputId": { "type": "string", "minLength": 1 },
          "format": { "type": "string", "enum": ["SWIFT_MT", "ISO_20022"] },
          "label": { "type": "string", "minLength": 1 },
          "messageIdentity": { "type": "string", "minLength": 1 },
          "mediaType": { "type": "string", "enum": ["application/json"] },
          "document": { "type": "object", "additionalProperties": true }
        }
      },
      "ResolutionPageExecutionResult": {
        "type": "object",
        "required": [
          "definitionId",
          "definitionVersion",
          "scenarioId",
          "fixtureBindingId",
          "outcome",
          "payloadGenerated",
          "confirmedResolutionCreated",
          "repairQueueCreated",
          "nvrOutcome",
          "fields",
          "outputs",
          "evidence"
        ],
        "properties": {
          "definitionId": { "type": "string" },
          "definitionVersion": { "type": "string" },
          "scenarioId": { "type": "string" },
          "fixtureBindingId": { "type": "string" },
          "outcome": {
            "enum": [
              "RESOLVED",
              "NOT_REQUIRED",
              "NO_ELIGIBLE_SSI",
              "VALIDATION_REJECTED",
              "REFERENCE_ONLY"
            ]
          },
          "reasonCode": { "type": "string" },
          "payloadGenerated": { "type": "boolean" },
          "confirmedResolutionCreated": { "type": "boolean" },
          "repairQueueCreated": { "type": "boolean" },
          "nvrOutcome": {
            "$ref": "#/components/schemas/PageParameterCompatibleNvrOutcome"
          },
          "fields": { "type": "array", "items": { "type": "object" } },
          "outputs": {
            "type": "array",
            "description": "Server-generated output representations from the same canonical resolution. Payment MT2/pacs.009 success returns SWIFT_MT and ISO_20022; clients do not convert them.",
            "items": {
              "$ref": "#/components/schemas/ResolutionPageGeneratedOutput"
            }
          },
          "evidence": {
            "type": "object",
            "required": [
              "correlationId",
              "owner",
              "action",
              "executorIdentity",
              "requestSha256",
              "responseSha256",
              "ruleIds"
            ],
            "properties": {
              "correlationId": { "type": "string" },
              "owner": {
                "enum": ["SSI_FIELD_RESOLUTION_API", "UPSTREAM_FIN_VALIDATOR"]
              },
              "action": {
                "enum": ["RESOLVE_SSI", "VALIDATE_FIN", "PREVIEW_REFERENCE"]
              },
              "executorIdentity": { "type": "string" },
              "requestSha256": { "type": "string" },
              "responseSha256": { "type": "string" },
              "ruleIds": { "type": "array", "items": { "type": "string" } }
            }
          }
        }
      },
      "Bic": {
        "type": "string",
        "pattern": "^[A-Z0-9]{4}[A-Z]{2}[A-Z0-9]{2}([A-Z0-9]{3})?$",
        "minLength": 8,
        "maxLength": 11,
        "description": "ISO 9362 BIC syntax only; directory existence requires licensed Bank/SwiftRef data."
      },
      "Currency": {
        "type": "string",
        "pattern": "^[A-Z]{3}$",
        "description": "Must also exist in Currency service ISO 4217 data."
      },
...
    }
  }
}